Share artwork rendering and layout between help and init
Validation: typecheck, lint, 17 artwork tests and 960 CLI tests pass. Full engine suite has 843 passing tests and the existing local clack consent re-prompt timeout.
